run-test: Move run-test-jvmti-java to CTS directory
run-test-jvmti-java will also now build against the android SDK,
which means it takes on a framework dependency.

Move the 'run-test-jvmti-java' make target from
art/test to cts/hostsidetests/jvmti/run-tests to avoid having
the ART buildbots compile it (which do not include the framework
in their manifest).

(It also prevents an 'mma' in ART directory from having to build the framework).

Bug: 63390896
Change-Id: I17b71d07559d0ef6f757f31369bd87b2d6e666d0