Include more cases in ImageWriter and ImageSpace
Add a reference visitor for object in image writer to prune not only classes
but also instances referenced. Otherwise an object storing a ref to a pruned
class will not pruned.

Add a check for ClassClass case when fixing up object pointer address in
ImaceSpace to fixing dependencies before fixing VTables.

Tests may not cover the new logic, but they will be tested when clinit is
supported for app image later.

Test: make test-art-host-j64
Change-Id: I8318cdb553eff8e56a30493124f859419f20e23f