Include more cases in ImageWriter
Add a reference visitor for object in image writer to prune not only classes
but also instances referenced. Otherwise an object storing a ref to a pruned
class will not pruned.

Prune a class anyway if it contains any field that is an instance of Method
and Constructor class because reflection is forbidden.

Tests may not cover the new logic, but they will be tested when clinit is
supported for app image later.

Test: make test-art-host-j64
Change-Id: I8318cdb553eff8e56a30493124f859419f20e23f