Update patch set 32
Patch Set 32:

FYI we're still exploring some potential issues with the test cases [see the slide deck discussion] so I'm not sure if current semantics are 100% solid.

---

As for the test code, would it make more sense to write them with some minimal smali/jasmin? I find it quite difficult to follow with 3+ partial compilations (and that's after trying to write several test cases out myself, let alone reading someone else's test cases).

We should already have jasmin available in the platform so it should just be a very simple matter of adding 'jasmin' support to default-build.

(just 2 compilations or less is probably fine) also majority of code could still be written in .java, only changing parts that absolutely need to be in bytecode.

Patch-set: 32