Verify the class loader context when loading oat files
Previously, the oat_file_manager would expect and perform validation on a
simple classpath: a list of dex files separated by ':'.

This is no longer enough since the oat file may encode a chain of class
loaders now. The CL moves the validation logic in ClassLoaderContext and
extends it to verify the complete chain of class loaders.

Test: m test-art-host
Bug: 38138251
Change-Id: I8ac9c65db1a14909aaecb04fa7a7115ddedc673f