Constrain the class memory usage in app image
Implemented a constraint for the space allowed per class in app image.
The default space allowed is 10kb per class, including field and arrays
and all instances created within the <clinit>, either stored in app
image or not.

The memory is counted by tracing the call to instructions INSTANCE and
NEW_ARRAY and NEW_FILLED_ARRAY, limitation is checked before space is
actually allocated. Because the initialization is bond with transaction,
the space usage is stored in transaction.

Testcase 660-clinit include a test case for this situation, which try to
create a large array in clinit but stopped.

Test: make test-art-host -j64
Change-Id: I5f440a12108b9845daa460f98a07483890a7aa9a