Assume >= SSE4.2/popcount for runtime built on x86/x86_64.
Rationale:
Move ART towards the right millenium and assume that when
a runtime is build for x86 and x86_64, it at least supports
SSE4.2 (128-bit SIMD) and popcount. Since this implies the
runtime will find the right flags through FromCppDefines(),
it enables taking advantage of these features, e.g. by means
of SIMDization.

Bug: 63585550

Test: test-art-host and test-art-target (FUGU)

Change-Id: I0f472254e20fe80a43b30ad06734a7ff5c14fa73