Change method lookup to follow JLS instead of JVMS.
If method lookup finds a method that is not inherited in the
superclass chain, we search the "copied" methods for another
method that is inherited from an interface and only return
the former when the search in "copied" methods fails. This
is a divergence from the RI but follows the semantics
implied by JLS (see 8.4.8 for inherited methods) and javac.

Test: m test-art-host-gtest
Test: testrunner.py --host
Test: testrunner.py --host --interp-ac
Test: Nexus 6P boots.
Test: testrunner.py --target
Bug: 62855082
Change-Id: I250f51f10d6e0c29b5955677348be058fe1a35df