Update patch set 10
Patch Set 10:

I would like to see some tests of classes that are checked to be *not* pre-initialized [by the app image], then in subsequent CLs you could relax that test as more things are pre-initialized.

It will also make it more obvious what each CL is actually accomplishing.

Patch-set: 10