ART: Try to statically evaluate some HIf in DCE.
For two HIfs with the same HCondition try to statically evaluate
the condition value for the second_if and transform it to an
unconditional branch using dominance information. Example:

  if (cond) {
    ...
    if (cond) {
      ...
    }
    ...
  }

The patch is a prerequisite step for "Loop peeling to eliminate
invariant exits" however it brings some value on its own with
a tiny code size reduction in boot-framework.oat (-8Kb).

Test: 548-checker-inlining-and-dce, test-art-target, test-art-host.
Change-Id: Ifbe45097dc2b5f098176fa1a1d023ea90b76d396