Make test 924 less flaky.
GetThreadState on threads blocked on monitors can be affected by GC
and other internals. This is because internal suspension of threads
marks them as Runnable. To make this test not observe this as much we
make it call GetThreadState multiple times and use the most common
result as the threads current state.

Bug: 62117833
Test: stress --cpu 64 &
      while ./test/run-test --host 924; do ; done

Change-Id: Iaa38cb58386416a72a6bb3c26ed90f4b8b623d1b