Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> (5 comments)
> 
> I'm not very happy with this change. You're polluting regular paths with a transaction check. I'd rather you consider alternatives.
> 
> Also, this CL should be split. Letting a Transaction know about the *Class* it initializes (not an Object*) is in know way inherent to Nested Transactions.

This makes a good point. I thinked a while and figured out that another reason we decide to use nested transactions instead of a list of classes in one transaction is, we want to partialy submitted a transaction. If another class is initialized within a transaction, called A, and is totally clear with A, the result should be committed. And if they have some inheretance relation, it aborts.

I will adjust the rest sanitization stuff, seperated the nest transaction and transaction awareness, and is there any other problems?

Patch-set: 11
Reviewer: Gerrit User 1167013 <1167013@85c56323-6fa9-3386-8a01-6480fb634889>