Implemented Nested Transaction
Transaction can be created recursively and every transaction knows which
class it is created for.

The goal of implementing nested transaction is to let every class
initialization (especially execution of class initializer) associates
with one transaction which provides the ability to track whether field
accesses are valid. To achieve this goal, we implemented a subclass for
ClassLinker called AotClassLinker, which will be instantiated instead of
ClassLinker under AOT compiler.

All invocations of InitializeClass happens at AOT compiler will be
wrapped with creating and destorying transactions.

Transactions will be paused when rollbacking, otherwise the memory will
not be rollbacked properly.

The use the transaction is changed and all usage are updated,
corresponding with native tests mentioning transaction, including
transaction_test and unstarted_runtime_test.

The validation rules will be implemented and explained in another CL.

Test: make test-art-host -j64

Change-Id: If53d3ee3231c337a9ea917f5b885c173917765de