Use mmapped boot image intern table for PIC app HLoadString.
Implement new HLoadString load kind for boot image strings
referenced by PIC-compiled apps (i.e. prebuilts) that uses
PC-relative load from a boot image InternTable mmapped into
the apps .bss. This reduces the size of the PIC prebuilts
that reference boot image strings compared to the kBssEntry
as we can completely avoid the slow path and stack map.

We separate the InternedStrings and ClassTable sections of
the image (.art) file from the rest, aligning the start of
the InternedStrings section to a page boundary. This may
actually increase the size of the image file by a page but
it also allows mprotecting() these tables as read-only.
The ClassTable section is included in anticipation of a
similar load kind for HLoadClass.

To simplify the recognition of different parts of the oat
.bss, we remove the DexCache arrays. Their presence in the
.bss was no longer necessary and it doesn't really matter
for memory usage whether they are in the .bss or in the
LinearAlloc. This also removes a lot of unnecessary code.

Prebuilt services.odex for aosp_angler-userdebug (arm64):
  - before: 20862776
  - after: 20308512 (-541KiB)
Note that 92KiB savings could have been achieved by simply
avoiding the read barrier, similar to the HLoadClass flag
IsInBootImage(). Such flag is now unnecessary.

Test: m test-art-host-gtest
Test: testrunner.py --host
Test: testrunner.py --host --pictest
Test: testrunner.py --target on Nexus 6P.
Test: testrunner.py --target --pictest on Nexus 6P.
Bug: 31951624
Change-Id: I5f2bf1fc0bb36a8483244317cfdfa69e192ef6c5