Runtime validation for <clinit> results
Add a global flag sotred in ClassLinker indicating whether boot space is
compromised -- final fields are modified by users through unexpected
techniques.

Add a special status, kStatusPreInitialized, for classes that
pre-initialized by compiler and identified that rely on boot image's
final static fields.

Combining those two new information gathered, we can check whether boot
image is compromised when we need to initialize a class marked as
PreInitialized, if so, re-run the <clinit>, otherwise, promote the
status to kStatusInitialized.

Add a test for this situation, modifying the ArrayList's final
static fieds through reflection, and see whether the <clinit> is re-ran
for classes reply on that field and initialized later than the
modification wad made.

Bug: 62887449
Test: make test-art-host -j64

Change-Id: Ic39af54a309d7dd28f507a50570e7d8715877027