MIPS: Don't use nop (0x00000000) for entrypoints padding
When ClassLinker tries to set the entrypoint for a native method,
OatFile::OatMethod::GetQuickCode() returns nullptr. This is done
because art_quick_generic_jni_trampoline is treated as OatMethod and
it is trying to get its size by loading word from address just prior
start of the entrypoint. There is a "nop" instruction used for padding
which is encoded as 0x00000000.

"or $zero, $zero, $zero" instruction will be used as padding to avoid
this problem. This instruction is encoded as 0x00000025.

This fixes two test failures:
* 945-obsolete-native
* runtime_callbacks_test

Note: All tests are executed in QEMU (MIPS64R6 and MIPS32R6) and on
      ci20 board (MIPS32R2).

Test: mma test-art-target-gtest
Test: ./testrunner.py --target --optimizing

Change-Id: I936d301be2b4abaa2fede721e6d05fa14310d01d