Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> @VMarko: Any comment on why you kept around GetInterfaceMethod at all? It seems to be essentially the same as GetClassMethod just slightly less general (and missing some DCHECKs like the one here).

We chatted offline about this.

The root cause for this seems to be the "abuse" on FindClassMethod in order to provide better error reporting. Igor and you suggested that's only needed for the Object case. So maybe we can special case it? I did not perform an in depth review of vladimir's change so others who did might want to jump in.

As part of the fix please don't forget to 
- improve the docs with rationale
- add another tests case in the vadimir's suite.

Patch-set: 3
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>