Update patch set 7
Patch Set 7: Code-Review+2

Looks good to me.

The only way to make this "prettier" (and it's probably more a side-grade) would be to put the few hard failures into a lambda, so that you can use an early return. But not sure that's really worth it.

Patch-set: 7
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2