Update patch set 5
Patch Set 5:

(3 comments)

Thanks for the comments guys.

I don't know enough about the behavior of the HandleCache and how accessible we want that to be. Maybe we can keep that independent of this change. I would be happy to do follow-up changes if we agree on how to deal with it. For now, I have made it private again in ReferenceTypePropagation.

Patch-set: 5
Reviewer: Gerrit User 1095168 <1095168@85c56323-6fa9-3386-8a01-6480fb634889>