Update patch set 1
Patch Set 1:

> (1 comment)

1. HInstruction is not trivially-copyable because it has a virtual destructor. Anyway, my point is that those tricks are somewhere close to the undefined/unspecified/implementation defined areas.

2. Yes, for DCHECK probably yes, and for the current state of the nodes.h maybe as well. However if some instruction will decide to increase a static variable in the contructor (or similar, some sideeffects), we'll have problems. In general I was just thinking about this as going behind C++ back and whether it is good design perspective.

So the cost of keeping this as it is - just an extra function definition in each instruction for getting more secure, C++ friendly and less error-prone code?

Anyway if you have a strong opinion on this I can rewrite it.

Patch-set: 1
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>