Update patch set 2
Patch Set 2:

> > > (1 comment)
 > >
 > > 1. HInstruction is not trivially-copyable because it has a
 > virtual
 > > destructor. Anyway, my point is that those tricks are somewhere
 > > close to the undefined/unspecified/implementation defined areas.
 > >
 > > 2. Yes, for DCHECK probably yes, and for the current state of the
 > > nodes.h maybe as well. However if some instruction will decide to
 > > increase a static variable in the contructor (or similar, some
 > > sideeffects), we'll have problems. In general I was just thinking
 > > about this as going behind C++ back and whether it is good design
 > > perspective.
 > >
 > > So the cost of keeping this as it is - just an extra function
 > > definition in each instruction for getting more secure, C++
 > > friendly and less error-prone code?
 > >
 > > Anyway if you have a strong opinion on this I can rewrite it.
 > 
 > I share the C++ semantics concern. But we need something to make
 > this robust. How about doing a memcmp after the cloning?

1. About robustness: which exactly problem do we want to solve: the redefinitions of virtual functions or a fact that for newly defined instruction we will have to bother and implement cloning related methods?

I could think of some other solutions like having a big switch-like function to call the constructors (to avoid virtual functions and changes in every instruction) - is it better than the current one?


2. I am sorry, Nicolas, I still think memcpy might be not a good option, as "If the objects are not TriviallyCopyable, the behavior of memcpy is not specified and may be undefined" (http://en.cppreference.com/w/cpp/string/byte/memcpy).

In this case "the target object is not alive. We destroyed it by reusing its storage, and haven't revitalized it by a constructor call."

"However, if a program ends the lifetime of an non-trivial object, it must ensure that a new object of the same type is constructed in-place (e.g. via placement new) before the destructor may be called implicitly, i.e. due to scope exit or exception for automatic objects, due to thread exit for thread-local objects, or due to program exit for static objects; otherwise the behavior is undefined." (http://en.cppreference.com/w/cpp/language/lifetime)

Some more info and comments in the thread (https://stackoverflow.com/questions/29777492/why-would-the-behavior-of-stdmemcpy-be-undefined-for-objects-that-are-not-triv)

Patch-set: 2
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>