Update patch set 1
Patch Set 1:

"Not seem to affect startup" Is there numbers somewhere? I recall maps launch is particularly susceptible to regressing when compiling fewer methods for speed profile.

Another area of concern is that if we have no samples for methods that don't have backwards branches, it might regress layout since we won't know if they were executed or not.

Patch-set: 1
CC: Gerrit User 1015378 <1015378@85c56323-6fa9-3386-8a01-6480fb634889>