Fix interaction of VMInit and ThreadStart events.
Real world agents require that one is able to create, and run, new
threads while the VMInit event is still being executed. Further, these
require that ThreadStart events can occur concurrently with the VMInit
event. This CL enables this behavior and adds a test for the
interaction of these two events.

Test: ./test.py --host -j50
Bug: 62821960
Bug: 34415266

Change-Id: I305f1ce3f1df9bf5a7e33027e0724f5fbac5c0f1