Update patch set 3
Patch Set 3:

Please consider a much simpler test for this. As it is, I think it will be very hard for other members to understand it if it breaks.

In essence, you really want to test that given a profile (which you can hardcode with just a couple of methods), the layout is what you expect. Sounds like you could even achieve this already with oatdump.

Patch-set: 3
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>