Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I don't understand what this is fixing. Can you add a run-test?

This is covered by vm-tests. As there's a bug open to move them to be directly testable I will not duplicate the issue here.

Problem is compile-time access checks on non-resolved fields vs interpreter execution. Old behavior skipped the check and incorrectly marked the class as completely verified, not re-verifying and thus not finding the AC violation and thus not running in the AC interpreter and thus not throwing a runtime error.

Patch-set: 4
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>