Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

Looks like a safe solution to me.
I don't understand enough of the context to tell whether there is an alternative way to avoid this situation. Mathieu?
We could also try to be more aggressive and just assume that we already own the lock, and thus are good to go without acquiring it. Probably not important enough to bother.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1