Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Can you please elaborate on how you want the test to be made simpler? 
> 
> It was intentionally made future-proof, so as to avoid being brittle: if we add more MethodHotness flags, only the 1 place in gen.py needs to be updated. If we change layout algorithm, only the 1 sort function needs to be updated.
> 
> Verifying more combinations means we have a higher certainty for correctness.

For sure :) I was not saying you should cover less.

> >  In particular I find it hard to follow the script the generates the java code and to match that against what the profile contains.
> 
> Can you clarify please?

The profile does not have a "GENERATED" comment so I thought you wrote it by hand.

> m_a$hot$startup$poststartup
> 
> - means it's a method with (hot|startup|poststartup) bits set.
> - m_a$$$ means it's a method with no bits set.
> - in general: m_a$X$YZ$ has (X|Y|Z) set.
> ----------
> 
> There's no reason to look at gen.py at all if the test is failing, it never gets executed by the build/run. It's only there if we change underlying layout implementation to avoid having to rewrite test.

We'll still need to look at gen.py if the test is failing due a generation issue.

Also, the relation between all the *expected* files is not clear. This is the only test having multiple expectations and they should be better documented in the info files.

Maybe better docs, as Andreas said, will improve things but as it is the script looks a bit overkill and long term harder to maintain than a plain java + profile file (it has more lines of code to read and understand than the code it generates :) ).

Patch-set: 3
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>