Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This seems like this should've included an example in the commit since we can't tell what this is actually fixing w.r.t to JLS semantics.

It's not fixing JLS semantics, really. It's one of those binary-compat cases, where I'm just trying to align more with the RI on whether we're pessimistic or not.

Assume:

 package a;
 public class A {
 }

 package b;
 class B {
   void test(A a) { }
 }

and an incompatible change to make A non-public. Then there will be no runtime-visible effect (there's no access to "a," so there's no IllegalAccessError for just calling test).

There will still be no runtime-visible effect with this commit. However, we will also no longer send B slow-path.

Patch-set: 3