Update patch set 1
Patch Set 1:

I'm torn between this and just using Atomic<pid_t> and pid_t as the owner type. That seems cleaner, but this fails more conspicuously of pid_t where somehow too large to be atomically updatable.

Posix says pid_t's are signed integers, which should be fine.

Are there any hazards to changing the size of an ART Mutex?

Owners are sometimes compared to -1 or 0. Any reason for the -1 comparison?

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>