Lower PLOG(ERROR) to PLOG(WARNING) if unable to find realpath
We were emitting a PLOG(ERROR) from class_loader_context if we are
unable to find the dex-files realpath. This is an issue because
InMemoryDexClassLoader and some --redefine-stress tests cause this
message to be printed, making the test fail.

This also brings it more into line with other logs in this file.

Test: ./test/run-test --host --jvmti-redefine-stress 071
Change-Id: Idc21bf89be02ca9263fbec7f28fdf036196ab58c