Update patch set 1
Patch Set 1:

I had to ask ngeoffray@ about this so I'll document here: the original submission was being used to see what the performance impact of a change was via the various automated build bots. When the performance data was available it was reverted. The ART change had "Temporarily merge this CL, then revert." but the libcore one did not.

Patch-set: 1
CC: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>