Update patch set 2
Patch Set 2:

> can we have a brief discussion in the Monday meeting about the
 > direction that all these tests are taking now (I am a bit worried
 > about no longer testing original Java)?

Ok, since we did not have time to get to this, let's just have the discussion here. I am just a bit worried that by rewriting into smali, we take something away from the great suite of Java tests, which has been used in the past to find issues with jack, back to dx, and now back to another front end. Perhaps it would be better to keep to keep the original java tests too (blacklist them for where it does not work), and simply add the smali equivalents to the side. That way our test suite is only getting more elaborate, not restrictive.

Thoughts?

Patch-set: 2