Update patch set 2
Patch Set 2:

> Thoughts?

Ah, our messages crossed. 

I see the argument for avoiding patterns that are really way too specific, but I just wanted to start a discussion on replacing java with smali. If this proliferates, we lose the ability to test other FE in the future quickly. If this will be just a few cases, I won't object (too strongly ;-)). 

Just wanted to sync our frequencies before proceeding...

Patch-set: 2