Update patch set 2
Patch Set 2:

> (Don't really understand why the image writer is a "linker" ...)

Well, it kind of links the heap image. But I would not object to moving it out of linker/ if you think it's confusing.

 > (Overall looks good to me. Should we consider a linker library?)

I was really struggling with moving the ElfWriter to dex2oat/. I eventually succeeded with an extra static library, libart-dex2oat. I do not want to increase the number of .so files shipped to the device.

Patch-set: 2
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>