Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What's the plan going forward? I think it would be OK to enable the handler on devices for just SIGILL (with just re-raising so debuggerd gets to do its work, too), which would be an "enhancement" of the previous CL.

This CL was motivated by not being able to repro this problem at all locally. However, a repro happened today and I've had a thought on how it might be possible to increase the repro rate (would be ecstatic with 1 an hour). If this is correct, it might be good have the enhancement but this can follow.

Patch-set: 1
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0