Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > My question about unnecessary code complexity for ONLY_USE_SYSTEM_OAT_FILES
>  > has not been answered. No tests, no benchmarks, only added code
>  > complexity and maintenance burden to see here.
> 
> Do you mind repeating your question here, and pointing to the code you feel is too complex?

If ART was the point enforcing the policy, then I'd understand adding code to ART. But it will be selinux enforcing the policy. So I don't understand how you want to add *any* code, because any code now becomes a maintenance burden and a burden of understanding.

So it's the standard questions: Why should the runtime be doing this, when we're not trusted to do it right, anyways (hey, we're not the bootloader or the kernel, so what do *we* know?)? What is the goal of ART, what is the goal of other parts of the platform, and why should the complexity live with us?

Put a selinux dontaudit (so there's no logspam) rule in that effectively hides odex files in /data, and the OatFileAssistant will transparently not find those files anymore and thus not load them. Except for "dontaudit," this is what will happen, anyways. So why add any code to ART?

Turn of dex2oat (which Runtime already has as an option), and you're done.

So we'll have to have zero understanding of some tied-down security-theater mode that the platform wants for specific apps.

Patch-set: 3
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>