Update patch set 4
Patch Set 4:

> Yes, I understand that :) But it's not really lse's fault that it's
 > broken :)

This CL does *not* set a precedent for removing one opt in favor of a new one. Both LSE and VEC are established optimizations by now and have both been actively worked on while the other was there. It is only recently that the unfortunate interaction between the two was found (where, objectively, LSE is at fault for not looking at SIMD l/s). I think it is fair to say what we all have been at the receiving and sending sides of such interactions before.

This CL *does* set a precedent for not leaving the mainline in a broken state while a solution is being worked on. I would have been happy to do that fully, but did not want to get in the way of Xueliang Zhong's work.

Patch-set: 4
Label: Presubmit-Ready=0