Update patch set 3
Patch Set 3:

Ithink we should move this discussion to a design doc. CL comments is not a good place for designing this.

IMO:
1) we can and probably should try to enforce this with selinux policies as well and I think nnk@ or jeffv@  should be consulted if that's doable. Thinking more about it, there might be some limitations to what is achievable. I think that should be the starting point anyway. I assume that the desire is to block any code, including *.so files. So the code in art will not suffice.

2) regardless of 1) I think it would be beneficial to enforce this at runtime as well (even if only as debug checks). Why? I think it's easier to reason.

It's true that it would add complexity to the runtime  but it will also make clear what is allowed and what not. The logic of what oat file is the best is already complex and have slight different on  host vs device.

With an SElinux only solution we just add a new implicit dimension to the oat decision. Implicit because we will not know why it failed. And that will make issues hard to investigate. I prefer the explicit complexity to the implicit one.

A few questions :
1) what is the strategy for updated-able system priv apps?
2) what about *.so files?
3) was there any eval done on the performance impact? What level of regressions are acceptable?

Patch-set: 3
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>